+++ packages/platforms/browser/lib/auto-instrumentation/route-change-plugin.ts
@@ -1,6 +1,5 @@
-import { type SpanOptionSchema, isString, coreSpanOptionSchema, validateSpanOptions, type InternalConfiguration, type Plugin, type SpanFactory } from '@bugsnag/core-performance'
+import { coreSpanOptionSchema, isString, validateSpanOptions, type InternalConfiguration, type Plugin, type SpanFactory, type SpanOptionSchema } from '@bugsnag/core-performance'
 import { type BrowserConfiguration } from '../config'
-import getAbsoluteUrl from '../request-tracker/url-helpers'
 import { type RouteChangeSpanOptions } from '../routing-provider'
 
 // exclude isFirstClass from the route change option schema
@@ -9,11 +8,6 @@ const routeChangeSpanOptionSchema: SpanOptionSchema = {
   startTime,
   parentContext,
   makeCurrentContext,
-  route: {
-    getDefaultValue: (value) => String(value),
-    message: 'should be a string',
-    validate: isString
-  },
   trigger: {
     getDefaultValue: (value) => String(value),
     message: 'should be a string',
@@ -22,7 +16,7 @@ const routeChangeSpanOptionSchema: SpanOptionSchema = {
 }
 
 interface InternalRouteChangeSpanOptions extends RouteChangeSpanOptions {
-  route: string
+  url: URL
   trigger: string
 }
 
@@ -38,11 +32,30 @@ export class RouteChangePlugin implements Plugin<BrowserConfiguration> {
 
     let previousRoute = configuration.routingProvider.resolveRoute(new URL(this.location.href))
 
-    configuration.routingProvider.listenForRouteChanges((route, trigger, options) => {
+    configuration.routingProvider.listenForRouteChanges((url, trigger, options) => {
+      let absoluteUrl
+
+      if (url instanceof URL) {
+        absoluteUrl = url
+      } else {
+        try {
+          const stringUrl = String(url)
+          absoluteUrl = new URL(stringUrl)
+        } catch (err) {
+          configuration.logger.warn('Invalid span options\n  - url should be a URL')
+
+          return {
+            id: '',
+            traceId: '',
+            isValid: () => false,
+            end: () => {}
+          }
+        }
+      }
+
       // create internal options for validation
       const routeChangeSpanOptions = {
         ...options,
-        route,
         trigger
       }
 
@@ -53,15 +66,15 @@ export class RouteChangePlugin implements Plugin<BrowserConfiguration> {
         configuration.logger
       )
 
-      // update the span name using the validated route
-      cleanOptions.name += cleanOptions.options.route
-      const span = this.spanFactory.startSpan(cleanOptions.name, cleanOptions.options)
+      const route = configuration.routingProvider.resolveRoute(absoluteUrl)
 
-      const url = getAbsoluteUrl(cleanOptions.options.route, this.document.baseURI)
+      // update the span name using the validated route
+      cleanOptions.name += route
+      const span = this.spanFactory.startSpan(cleanOptions.name, cleanOptions.options)
 
       span.setAttribute('bugsnag.span.category', 'route_change')
       span.setAttribute('bugsnag.browser.page.route', route)
-      span.setAttribute('bugsnag.browser.page.url', url)
+      span.setAttribute('bugsnag.browser.page.url', url.toString())
       span.setAttribute('bugsnag.browser.page.previous_route', previousRoute)
       span.setAttribute('bugsnag.browser.page.route_change.trigger', cleanOptions.options.trigger)