+++ test/browser/features/support/maze-config.rb
@@ -1,11 +1,15 @@
 require_relative '../lib/browser'
+require_relative '../lib/build-mode'
 require_relative '../lib/url-generator'
 
+$build_mode = BuildMode.new
 $browser = Browser.new(Maze.config.browser)
 
 Maze.hooks.before_all do
   Maze.config.document_server_root = File.realpath("#{__dir__}/../fixtures/packages")
   Maze.config.enforce_bugsnag_integrity = false
+
+  $logger.info("Running in #{$build_mode.mode_name} mode")
 end
 
 Maze.hooks.before do